Move check_for_type_support() to its own module
