+++ rclpy/rclpy/node.py
@@ -20,7 +20,6 @@ from rclpy.client import Client
 from rclpy.clock import ROSClock
 from rclpy.constants import S_TO_NS
 from rclpy.exceptions import NotInitializedException
-from rclpy.exceptions import NoTypeSupportImportedException
 from rclpy.expand_topic_name import expand_topic_name
 from rclpy.guard_condition import GuardCondition
 from rclpy.impl.implementation_singleton import rclpy_implementation as _rclpy
@@ -34,6 +33,7 @@ from rclpy.service import Service
 from rclpy.subscription import Subscription
 from rclpy.time_source import TimeSource
 from rclpy.timer import WallTimer
+from rclpy.type_support import check_for_type_support
 from rclpy.utilities import get_default_context
 from rclpy.validate_full_topic_name import validate_full_topic_name
 from rclpy.validate_namespace import validate_namespace
@@ -41,22 +41,6 @@ from rclpy.validate_node_name import validate_node_name
 from rclpy.validate_topic_name import validate_topic_name
 
 HIDDEN_NODE_PREFIX = '_'
-
-
-def check_for_type_support(msg_type):
-    try:
-        ts = msg_type.__class__._TYPE_SUPPORT
-    except AttributeError as e:
-        e.args = (
-            e.args[0] +
-            ' This might be a ROS 1 message type but it should be a ROS 2 message type.'
-            ' Make sure to source your ROS 2 workspace after your ROS 1 workspace.',
-            *e.args[1:])
-        raise
-    if ts is None:
-        msg_type.__class__.__import_type_support__()
-    if msg_type.__class__._TYPE_SUPPORT is None:
-        raise NoTypeSupportImportedException()
 
 
 class Node:
